Cache the DataPtrs in CUDAFuture
CUDAFuture needs to inspect the value it contains in order to first determine what devices its tensors reside on (so that it can record events on those devices), and then to record these tensors with the caching allocator when they are used in other streams. Extracting data ptrs can become somewhat expensive (especially if we resort to using the pickler to do that), hence it's probably a good idea to cache the result the first time we compute it.

Differential Revision: [D25303486](https://our.internmc.facebook.com/intern/diff/D25303486/)

ghstack-source-id: 117759986
Pull Request resolved: https://github.com/pytorch/pytorch/pull/48788